src/requirements.txt
# Core analytics dependencies
numpy>=1.26.0
pandas==2.2.2
scikit-learn>=1.3.0

# API service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0

# Dashboard
streamlit==1.28.1
plotly==5.17.0

# Database support
psycopg2-binary==2.9.9
sqlalchemy==2.0.31

# Cloud support (optional)
google-cloud-bigquery>=3.13.0
db-dtypes>=1.0.0